curated_hub/types.py to types.py
