+++ tests/unit/sagemaker/jumpstart/curated_hub/test_utils.py
@@ -13,9 +13,9 @@
 from __future__ import absolute_import
 
 from unittest.mock import Mock
-from sagemaker.jumpstart.curated_hub import utils
+from sagemaker.jumpstart.types import HubArnExtractedInfo
 from sagemaker.jumpstart.constants import JUMPSTART_DEFAULT_REGION_NAME
-from sagemaker.jumpstart.curated_hub.types import HubArnExtractedInfo
+from sagemaker.jumpstart.curated_hub import utils
 
 
 def test_get_info_from_hub_resource_arn():
